pagina_streamlit/requirements.txt
numpy
opencv-python-headless
ultralytics
streamlit-webrtc
av
pillow
gdown